Step by step instructions
First draft of a step-by-step manual about how to contribute using GitHub. Each paragraph could correspond to one step on the workflow image (first contributor role). We could give only one procedure to do any type of changes or make a separate paragraphs for each type of change. Images of the icons/button and additional links need to be added.