Updated file naming
Wrote about what to consider when naming files and added also file naming tips & examples of fields and file names. However, I'm not sure about the hierarchy with the subtopics, whether is okay: "### Example elements to include in the name" and "Examples of good file names"? Also should the tips be numbered or with bullet points?