Update nels_assembly.md
Using a country keyword for a tool assembly might create undesired results when courses for other independent national resources/tools are registered into TeSS. 

With the current suggestion some results of the query get lost, but maybe the annotations on TeSS should be improved (e.g. including explicitly NeLS as a keyword or similar).